refactor(Auth): Auth 클래스 Interface로 변경
객체지향하도록 Interface로 Controller가 접근하게 개선되었습니다.
